feat: make inliner a proper executor
