+++ packages/@code-like-a-carpenter/nx/executors/inliner/schema.d.ts
@@ -1,0 +1,21 @@
+/* eslint-disable */
+/**
+ * This file was automatically generated by json-schema-to-typescript.
+ * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
+ * and run json-schema-to-typescript to regenerate this file.
+ */
+
+export interface InlinerExecutor {
+  /**
+   * The name of the export to inline
+   */
+  exportName: string;
+  /**
+   * The file to inline
+   */
+  sourceFile: string;
+  /**
+   * The file to write the inlined content to
+   */
+  targetFile: string;
+}
