+++ packages/@code-like-a-carpenter/nx/package.json
@@ -3,6 +3,9 @@
   "main": "./src/index.ts",
   "executors": "./executors.json",
   "dependencies": {
+    "@babel/generator": "^7.23.6",
+    "@babel/template": "^7.23.9",
+    "@babel/types": "^7.23.9",
     "@nx/devkit": "17.2.8",
     "esbuild": "^0.20.0",
     "glob": "^10.3.10",
